backend/requirements.txt
fastapi
uvicorn[standard]
subconscious-sdk